Move cron jobs to a separate file
